Refactor `donnafin.commons.core.index` -> `..core.types`
In preparation for upcoming dollar value.
