+++ src/main/java/donnafin/logic/commands/DeleteCommand.java
@@ -5,7 +5,7 @@ import static java.util.Objects.requireNonNull;
 import java.util.List;
 
 import donnafin.commons.core.Messages;
-import donnafin.commons.core.index.Index;
+import donnafin.commons.core.types.Index;
 import donnafin.logic.commands.exceptions.CommandException;
 import donnafin.model.Model;
 import donnafin.model.person.Person;
